requirements/transifex.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asttokens==2.4.0
    # via transifex-python
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via transifex-python
future==0.18.3
    # via pyseeyou
gitdb==4.0.10
    # via gitpython
gitpython==3.1.35
    # via transifex-client
idna==3.4
    # via requests
parsimonious==0.10.0
    # via pyseeyou
pyseeyou==1.0.2
    # via transifex-python
python-slugify==4.0.1
    # via transifex-client
pytz==2023.3.post1
    # via transifex-python
regex==2023.8.8
    # via parsimonious
requests==2.31.0
    # via
    #   -r requirements/transifex.in
    #   transifex-client
    #   transifex-python
six==1.16.0
    # via
    #   asttokens
    #   transifex-client
smmap==5.0.0
    # via gitdb
text-unidecode==1.3
    # via python-slugify
toolz==0.12.0
    # via pyseeyou
transifex-client==0.14.4
    # via -r requirements/transifex.in
transifex-python==3.4.0
    # via -r requirements/transifex.in
urllib3==1.26.16
    # via
    #   requests
    #   transifex-client
